Reorganize sharing components outside of the WorkPackage scope
