+++ app/components/_index.sass
@@ -1,5 +1,5 @@
-@import "work_packages/share/modal_body_component"
-@import "work_packages/share/invite_user_form_component"
+@import "shares/modal_body_component"
+@import "shares/invite_user_form_component"
 @import "work_packages/progress/modal_body_component"
 @import "open_project/common/attribute_component"
 @import "filter/filters_component"
