+++ app/components/shares/bulk_permission_button_component.html.erb
