+++ app/components/shares/bulk_selection_counter_component.html.erb
