+++ app/components/shares/concerns/authorization.rb
@@ -1,6 +1,8 @@
-#-- copyright
+# frozen_string_literal: true
+
+# -- copyright
 # OpenProject is an open source project management software.
-# Copyright (C) 2012-2024 the OpenProject GmbH
+# Copyright (C) 2023 the OpenProject GmbH
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License version 3.
@@ -24,17 +26,17 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #
 # See COPYRIGHT and LICENSE files for more details.
-#++
+# ++
 
-module WorkPackages
-  module Share
-    class ModalUpsaleComponent < ApplicationComponent
-      include ApplicationHelper
-      include OpTurbo::Streamable
-      include OpPrimer::ComponentHelpers
+module Shares
+  module Concerns
+    module Authorization
+      extend ActiveSupport::Concern
 
-      def self.wrapper_key
-        "work_package_share_list"
+      included do
+        def sharing_manageable?
+          User.current.allowed_in_project?(:share_work_packages, @work_package.project)
+        end
       end
     end
   end
