+++ app/components/shares/counter_component.html.erb
@@ -5,9 +5,9 @@
       # I'm able to manage shares if the only user that the work package is
       # currently shared is myself, since I'm not able to manage my own share.
       if sharing_manageable? && shared_with_anyone_else_other_than_myself?
-        render(WorkPackages::Share::BulkSelectionCounterComponent.new(count:))
+        render(Shares::BulkSelectionCounterComponent.new(count:))
       else
-        render(WorkPackages::Share::ShareCounterComponent.new(count:))
+        render(Shares::ShareCounterComponent.new(count:))
       end
     end
   end
