+++ app/components/shares/modal_body_component.sass
