+++ app/components/shares/share_counter_component.html.erb
