+++ app/components/shares/share_counter_component.rb
@@ -28,24 +28,16 @@
 # See COPYRIGHT and LICENSE files for more details.
 # ++
 
-module WorkPackages
-  module Share
-    module Concerns
-      module DisplayableRoles
-        def options
-          [
-            { label: I18n.t("work_package.sharing.permissions.edit"),
-              value: Role::BUILTIN_WORK_PACKAGE_EDITOR,
-              description: I18n.t("work_package.sharing.permissions.edit_description") },
-            { label: I18n.t("work_package.sharing.permissions.comment"),
-              value: Role::BUILTIN_WORK_PACKAGE_COMMENTER,
-              description: I18n.t("work_package.sharing.permissions.comment_description") },
-            { label: I18n.t("work_package.sharing.permissions.view"),
-              value: Role::BUILTIN_WORK_PACKAGE_VIEWER,
-              description: I18n.t("work_package.sharing.permissions.view_description") }
-          ]
-        end
-      end
+module Shares
+  class ShareCounterComponent < ApplicationComponent
+    def initialize(count:)
+      super
+
+      @count = count
     end
+
+    private
+
+    attr_reader :count
   end
 end
