+++ app/components/shares/share_row_component.rb
@@ -28,18 +28,51 @@
 # See COPYRIGHT and LICENSE files for more details.
 # ++
 
-module WorkPackages
-  module Share
-    class BulkSelectionCounterComponent < ApplicationComponent
-      def initialize(count:)
-        super
+module Shares
+  class ShareRowComponent < ApplicationComponent
+    include ApplicationHelper
+    include OpTurbo::Streamable
+    include OpPrimer::ComponentHelpers
+    include Shares::Concerns::Authorization
 
-        @count = count
+    def initialize(share:,
+                   container: nil)
+      super
+
+      @share = share
+      @work_package = share.entity
+      @principal = share.principal
+      @container = container
+    end
+
+    def wrapper_uniq_by
+      share.id
+    end
+
+    private
+
+    attr_reader :share, :work_package, :principal, :container
+
+    def share_editable?
+      @share_editable ||= User.current != share.principal && sharing_manageable?
+    end
+
+    def grid_css_classes
+      if sharing_manageable?
+        "op-share-wp-modal-body--user-row_manageable"
+      else
+        "op-share-wp-modal-body--user-row"
       end
+    end
 
-      private
-
-      attr_reader :count
+    def select_share_checkbox_options
+      {
+        name: "share_ids",
+        value: share.id,
+        scheme: :array,
+        label: principal.name,
+        visually_hide_label: true
+      }
     end
   end
 end
