+++ app/components/shares/user_details_component.html.erb
